// monorepo/apps/web/src/modules/warp/packages/client/features/form/types.ts
export type StateType = {
  formId?: string | null;
  formInvitationId?: string | null;
  formSubmissionId?: string | null;
  formFields: GetRenderFormFieldDetailsQuery["FormInvitation"][0]["Form"]["FormFields"];
  current: FormFieldWithChildrenType[];
  childFieldState?: Record<string, { value: any }>;
  answer: Record<
    string,
    {
      questionId: string;
      formFieldId: string;
      field: string;
      value: any;
    }
  >;
  isFormSubmitted?: boolean | null;
  warningRule?: boolean | null;
  isCarryForward?: boolean;
  Suggestions: GetRenderFormFieldDetailsQuery["FormInvitation"][0]["Suggestions"];
  isAIDataPointsAdded: boolean;
  isReviewer?: boolean;
  isMaker?: boolean;
  reviewerStatusMap?: Map<string, any>;
  commentCounts: Record<string, number>;
  formFieldMap?: Record<string, GetRenderFormFieldDetailsQuery["FormInvitation"][0]["Form"]["FormFields"][0]>;
};
// ...
export type DbAnswersType =
  GetRenderFormDetailsQuery["FormInvitation"][0]["FormSubmissions"][0]["Answers"];

export type DbFormFieldOptionsType =
  GetRenderFormFieldDetailsQuery["FormInvitation"][0]["Form"]["FormFields"];
export type DbInterim_RecommendationsType =
  GetRecommendationRenderFormDetailsQuery["FormInvitation"][0]["FormSubmissions"][0]["Interim_Answers"][0]["Interim_Recommendations"];
export type DbInterimAnswersType =
  GetRecommendationRenderFormDetailsQuery["FormInvitation"][0]["FormSubmissions"][0]["Interim_Answers"];
export type InterimAnswerStateType = {
  formFields: GetRenderFormFieldDetailsQuery["FormInvitation"][0]["Form"]["FormFields"];
  interim_answers: {
    formFieldId: string | null;
    data: any;
    isViewOnly: any;
    interim_recommendation: DbInterim_RecommendationsType;
    interim_answer: any;
  }[];
};

export type InterimAnswerActionsType = {
  init: (
    formFields: GetRenderFormDetailsQueryFormFieldType[],
    interim_answers: DbInterimAnswersType
  ) => void;
};
export type ActionsType = {
  init: (
    formId: string,
    formInvitationId: string,
    formSubmissionId: string,
    formFields: GetRenderFormDetailsQueryFormFieldType[],
    answers: DbAnswersType,
    isFormSubmitted: boolean,
    isCarryForward: boolean,
    Suggestions: GetRenderFormDetailsQuerySuggestionType,
    isAIDataPointsAdded: boolean,
    isReviewer: boolean,
    isMaker: boolean,
    reviewerStatusMap: Map<string, any>
  ) => void;
  setAnswer: (formFieldId: string, value: any) => void;
  setAnswers: (answers: Record<string, any>) => void;
  removeAnswer: (formField: string) => void;
  setChildFieldState: (name: string, value: any) => void;
  setFormFieldInterfaceOptions: (
    formFieldId: string,
    partialInterfaceOptions: any
  ) => void;
  setCommentCounts: (counts: Record<string, number>) => void;
};

export type FormFieldWithChildrenType = FormField & {
  children?: FormFieldWithChildrenType[];
};

export type FormFieldOptionsType =
  | "displayOptions"
  | "interfaceOptions"
  | "fieldOptions";

export type InfoIconPropsType = {
  content: "";
  popoverProps?: PopoverProps;
  popoverDropdownProps?: PopoverDropdownProps;
};

export type FormFieldInterfaces = {
  default: FormFieldWithChildrenType;
  "select-toggle": {
    type: "boolean";
    fieldOptions: {
      label: string;
      enable: boolean;
      readonly: boolean;
      required: boolean;
    };
    interfaceOptions: {
      showLabel: boolean;
      title: string;
      size: string;
      onLabel: string;
      offLabel: string;
      isHeading?: boolean;
      headingSize?: string;
      showTitleDivider?: boolean;
      infoIconProps?: InfoIconPropsType;
      isAddcomment?: boolean;
      subtitle: string;
    };
    display: string;
    displayOptions: any;
    displayRules: RuleType<
      Pick<FormFieldInterfaces["select-toggle"], FormFieldOptionsType>
    >[];
    validationRules: ValidationRuleType[];
  };
// ...
};

export type FormFieldInterfaceKeysType = keyof FormFieldInterfaces;

export type FormFieldInterfaceType<T extends FormFieldInterfaceKeysType> = Omit<
  FormField,
  keyof FormFieldInterfaces[T]
> &
  FormFieldInterfaces[T];

export type RuleType<TOptions = any> = {
  name: string;
  rule: string;
} & TOptions;

export type ValidationRuleType<T = any> = RuleType<
  T & {
    message: string;
  }
>;
// ...
export type GetRenderFormDetailsQueryFormFieldType =
  GetRenderFormFieldDetailsQuery["FormInvitation"][0]["Form"]["FormFields"][0];

export type GetRenderFormDetailsQuerySuggestionType =
  GetRenderFormFieldDetailsQuery["FormInvitation"][0]["Suggestions"];
// ...
export type warningmessageObjectType = {
  isWarningRule?: boolean;
  oldvalue?: string;
  newValue?: string;
  ratio?: string;
  warningmessage?: string;
  formfieldid?: string;
  Threshold_Deviation?: string;
  Deviation_differential?: string;
  questionid?: string;
  ispopupmessageremoved?: boolean;
  isFileUpload?: boolean;
  fileId?: string;
  invitationId?: string;
};

export type WarningMessageActionsType = {
  init: (WarningRuleFields: warningmessageObjectType[]) => void;
  removeWarningRuleFields: () => void;
};

export type WarningMessageStateType = {
  WarningRuleFields: warningmessageObjectType[];
};

export type GlobalMasterData = {
  data?: any;
  id?: string | null;
  platformId?: string | null;
  type?: string | null;
  __typename?: string | null;
} | null;

// export type RatingValidationType = {
//   data?: any;
//   id: string;
//   type?: string;
//   invitationId: string;
//   submissionId: string;
//   formFieldId: string;
//   created_at: any;
// };
export type RatingValidationStateType = {
  RatingValidation?: RaraValidationAndRating;
};

export type RatingValidationActionsType = {
  init: (value: RaraValidationAndRating) => void;
};